Update ASCIITable.xml
I have added the complete Extended ASCII to the table as described at http://ascii-code.com/.  Please note that if any of the >127 characters are entered in the value as HTML, like for example "&euro;" the Clarion toolbox stops working and is completely blank!  So I have left the values empty.  This has been tested with Clarion 10.0.11834.